requirements.txt
Django==1.11.1
django-js-reverse==0.7.3
django-polymorphic==1.2
django-widget-tweaks==1.4.1
djangorestframework==3.6.3
psycopg2==2.7.1
pytz==2017.2
xlrd==1.0.0
XlsxWriter==0.9.6
